refactor: splitt ChargenWizardExperience i steg og hook (#105)
- ChargenWizardExperience.tsx: 3770 → 138 linjer (under 150-grensen)
- chargenWizardHelpers.ts: typer, interfaces og rene hjelpefunksjoner
- useChargenWizardState.tsx: 35 state-variabler, derived values,
  effects og alle handlers
- Nye stegkomponenter: CivilizationStep, SocialClassStep,
  ProfessionStep, SkillAllocationStep, OtherSkillsStep,
  SpecializationsStep, SkillsTableStep, SpecializationTableStep,
  ReviewStep, FinalizeStep
- Eksporterte hjelpefunksjoner re-eksporteres via index.ts slik at
  ChargenWizard.test.ts-importkjeden forblir intakt
- 48 testfiler, 239 tester grønne
