+++ apps/web/src/features/chargen/index.ts
@@ -8,6 +8,6 @@ export {
   getSpecializationPurchaseState,
   getSpecializationRowMessages,
   getSkillRowMessages,
-} from "./ChargenWizardExperience";
+} from "./chargenWizardHelpers";
 export { FeedbackPanel } from "./components";
 export { ResolveStatsStep, StartStep, StatsStep } from "./steps";
