+++ apps/web/src/features/chargen/steps/index.ts
@@ -1,3 +1,13 @@
+export { CivilizationStep } from "./CivilizationStep";
+export { FinalizeStep } from "./FinalizeStep";
+export { OtherSkillsStep } from "./OtherSkillsStep";
+export { ProfessionStep } from "./ProfessionStep";
 export { ResolveStatsStep } from "./ResolveStatsStep";
+export { ReviewStep } from "./ReviewStep";
+export { SkillAllocationStep } from "./SkillAllocationStep";
+export { SkillsTableStep } from "./SkillsTableStep";
+export { SocialClassStep } from "./SocialClassStep";
+export { SpecializationTableStep } from "./SpecializationTableStep";
+export { SpecializationsStep } from "./SpecializationsStep";
 export { StartStep } from "./StartStep";
 export { StatsStep } from "./StatsStep";
